% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-hurdle-methods.R
\name{AIC.beezdemand_cp_hurdle}
\alias{AIC.beezdemand_cp_hurdle}
\title{AIC method for beezdemand_cp_hurdle}
\usage{
\method{AIC}{beezdemand_cp_hurdle}(object, ..., k = 2)
}
\arguments{
\item{object}{A beezdemand_cp_hurdle object}

\item{...}{Additional arguments (ignored)}

\item{k}{Penalty per parameter (default 2 for AIC)}
}
\description{
AIC method for beezdemand_cp_hurdle
}
