% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Demand Analyzer}
\usage{
analyze(adf = NULL, eq = NULL, nltype = NULL, k = NULL, q0st = 10,
  ast = 0.001, seetrace = FALSE, replfree = 0.01)
}
\arguments{
\item{adf}{A dataframe consisting of prices (x) and consumption values (y).}

\item{eq}{Character vector of length one. Accepts either "hs" for Hursh and Silberberg (2008) or "koff" for Koffarnus, Franck, Stein, and Bickel (2015). If "hs" and the first price (x) is 0, it will be replaced by replfree.}

\item{nltype}{Character vector of length one. "default" will use the nonlinear (weighted) least-squares estimate and Gauss-Newton algorithm (for more information, see ?nls). "lm" will fit using the Levenberg-Marquardt algorithm (for more information, see ?nlsLM in the minpack.lm package). "brute" will fit using a two-stage process. First, a grid of potential starting values between q0 and alpha is created; 1 value of q0 being the maximum consumption value in the given dataframe and 100 values of alpha ranging from 0.00000001 to 0.1. Since q0 is recycled, the residual sum of squares is evaluated in each combination of q0 and alpha. The combination of of q0 and alpha that result in the least residual sum of squares is subsequently used as starting values in the nonlinear least-squares fitting process. The port algorithm is used to allow for lower and upper bounds. Lower and upper bounds for q0 are 0.001 and 1.5 times the maximum consumption; lower and upper bounds for alpha are 0.0000000001 and 1. For all fitting methods, a maximum of 1000 iterations are used.}

\item{k}{A numeric vector of length one. Reflects the range of consumption in log10 units.}

\item{q0st}{A numeric vector of length one. Starting value used for q0 in both "default" and "lm" fitting methods. Default value is 10.}

\item{ast}{A numeric vector of length one. Starting value used for alpha in both "defualt" and "lm" fitting methods. Default value is 0.001}

\item{seetrace}{If TRUE, intermediate results may be shown. Default value is FALSE.}

\item{replfree}{If eq = "hs" and 0 is first price, 0 gets replaced by replfree. Default value is .01}
}
\value{
In most cases, an object of class 'nls' is returned. If there is an error in the fitting process, the convergence note (type = character) will be returned.
}
\description{
Analyzes a dataframe and returns the regression model.
}
\details{
This function takes a dataframe, consisting of prices (x) and consumption values (y), and analyzes it using nonlinear regression.
}
\author{
Brent Kaplan <bkaplan4@ku.edu>
}

