% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{annotation_logticks2}
\alias{annotation_logticks2}
\title{annotation_logticks2}
\usage{
annotation_logticks2(base = 10, sides = "bl", scaled = TRUE,
  short = unit(0.1, "cm"), mid = unit(0.2, "cm"), long = unit(0.3,
  "cm"), colour = "black", size = 0.5, linetype = 1, alpha = 1,
  data = data.frame(x = NA), color = NULL, ...)
}
\arguments{
\item{base}{base for drawing in scale}

\item{sides}{sides to draw, by default bottom and left}

\item{scaled}{true by default}

\item{short}{short tick settings}

\item{mid}{mid tick settings}

\item{long}{long tick settings}

\item{colour}{default to black colour}

\item{size}{size for labels}

\item{linetype}{default linetype}

\item{alpha}{default alpha level}

\item{data}{data to include}

\item{color}{colors to include}

\item{...}{additional arguments}
}
\value{
ggplot2 layer
}
\description{
Creates annotation layer
}
\details{
Inherit and extend layer for use in ggplot draw
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
