% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apt}
\alias{apt}
\title{Example alcohol purchase task data}
\format{A matrix with price in the first column and participant responses in subsequent columns. Participants were asked how many standard sized alcoholic beverages they would buy at various prices.}
\usage{
apt
}
\description{
A dataset containing alcohol purchase task data for a small number of participants
}
\keyword{datasets}

