% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{augment.beezdemand_tmb}
\alias{augment.beezdemand_tmb}
\title{Augment a beezdemand_tmb Model}
\usage{
\method{augment}{beezdemand_tmb}(x, newdata = NULL, ...)
}
\arguments{
\item{x}{A \code{beezdemand_tmb} object.}

\item{newdata}{Optional data frame.}

\item{...}{Additional arguments.}
}
\value{
A tibble with original data plus .fitted and .resid columns.
}
\description{
Augment a beezdemand_tmb Model
}
