\name{beezdemand-internal}
\alias{getSumOfSquaresExponential}
\alias{getSumOfSquaresExponentiated}
\alias{lseq}
\alias{minTicks}
\alias{trim.leading}
\alias{ExtractCoefs}
\alias{FitCurves.linear}
\alias{ExtractCoefs.linear}
\title{Internal Beezdemand Functions}
\description{
 Internal Beezdemand functions
}
\details{
These are not to be called by the user.
}
\keyword{ internal }
