% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-hurdle-methods.R
\name{BIC.beezdemand_cp_hurdle}
\alias{BIC.beezdemand_cp_hurdle}
\title{BIC method for beezdemand_cp_hurdle}
\usage{
\method{BIC}{beezdemand_cp_hurdle}(object, ...)
}
\arguments{
\item{object}{A beezdemand_cp_hurdle object}

\item{...}{Additional arguments (ignored)}
}
\description{
BIC method for beezdemand_cp_hurdle
}
