% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdle-methods.R
\name{BIC.beezdemand_hurdle}
\alias{BIC.beezdemand_hurdle}
\title{BIC for Hurdle Demand Model}
\usage{
\method{BIC}{beezdemand_hurdle}(object, ...)
}
\arguments{
\item{object}{A \code{beezdemand_hurdle} object.}

\item{...}{Additional arguments (unused).}
}
\value{
A numeric BIC value.
}
\description{
BIC for Hurdle Demand Model
}
