% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{calcAUC}
\alias{calcAUC}
\title{Calculate AUC}
\usage{
calcAUC(dfs, qmaxs = NULL)
}
\arguments{
\item{dfs}{A list of dataframes containing x (prices) and y (consumption).}

\item{qmaxs}{Required TRUE/FALSE. If TRUE, the maximum consumption value across all dataframes is used for normalization. If FALSE, the maximum consumption value for EACH dataframe is used for normalization.}
}
\value{
Numeric vector of AUC values.
}
\description{
Calculates Area Under the Curve
}
\details{
Calculates Area Under the Curve as described in Amlung, Yurasek, McCarty, MacKillop, & Murphy (2015).
}
\author{
Brent Kaplan <bkaplan4@ku.edu>
}

