% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change0}
\alias{change0}
\title{Change 0s}
\usage{
change0(mat, repl = NULL, replnum = 0.01, parshift = FALSE,
  prices = NULL)
}
\arguments{
\item{mat}{A matrix of consumption values.}

\item{repl}{The number of 0s to be replaced by replnum. Default is NULL (i.e., all 0s).}

\item{replnum}{Replacement number. Default value is 0.01.}

\item{parshift}{Whether to apply a parallel shift of replnum to all consumption and prices.}

\item{prices}{If parshift == TRUE, value is required.}
}
\value{
Matrix of transformed data. If parshift == TRUE, list of 2. First element is the transformed matrix. Second element is transformed price vector.
}
\description{
Changes 0 values or applies a parallel shift.
}
\details{
Takes a matrix of consumption values and replaces specified number of 0s with replacement value. Optionally, can apply a parallel shift to all consumption values and prices.
}
\author{
Brent Kaplan <bkaplan4@ku.edu>
}

