% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-methods.R
\name{coef.beezdemand_fixed_grouped}
\alias{coef.beezdemand_fixed_grouped}
\title{Extract Coefficients from Grouped Fixed-Effect Demand Fit}
\usage{
\method{coef}{beezdemand_fixed_grouped}(object, ...)
}
\arguments{
\item{object}{A beezdemand_fixed_grouped object}

\item{...}{Additional arguments passed to \code{coef.beezdemand_fixed()}}
}
\value{
A tibble with group columns prepended
}
\description{
Extract Coefficients from Grouped Fixed-Effect Demand Fit
}
