% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-hurdle-utils.R
\name{compare_cp_hurdle_models}
\alias{compare_cp_hurdle_models}
\title{Compare Two Cross-Price Hurdle Models}
\usage{
compare_cp_hurdle_models(model_full, model_reduced)
}
\arguments{
\item{model_full}{The full model (more parameters, e.g., 3-RE)}

\item{model_reduced}{The reduced model (fewer parameters, e.g., 2-RE)}
}
\value{
A list with class "cp_hurdle_comparison" containing test results
}
\description{
Performs a likelihood ratio test to compare two nested cross-price hurdle models.
Typically used to compare 2-RE and 3-RE models.
}
\examples{
\dontrun{
fit2 <- fit_cp_hurdle(data, "y", "x", "id", random_effects = c("zeros", "qalone"))
fit3 <- fit_cp_hurdle(data, "y", "x", "id", random_effects = c("zeros", "qalone", "I"))
compare_cp_hurdle_models(fit3, fit2)
}
}
\keyword{internal}
