% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint-hurdle.R
\name{compute_joint_start_values}
\alias{compute_joint_start_values}
\title{Compute Starting Values for Joint Hurdle Model}
\usage{
compute_joint_start_values(dat, k, epsilon, stream_counts)
}
\arguments{
\item{dat}{Prepared data frame}

\item{k}{Fixed k value (or NULL)}

\item{epsilon}{Epsilon value}

\item{stream_counts}{Named vector of stream counts}
}
\value{
Named list of starting values
}
\description{
Compute Starting Values for Joint Hurdle Model
}
\keyword{internal}
