% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint-hurdle.R
\name{compute_joint_start_values_saturated}
\alias{compute_joint_start_values_saturated}
\title{Compute Starting Values for Saturated Joint Hurdle Model}
\usage{
compute_joint_start_values_saturated(dat, k_start, epsilon, stream_counts)
}
\arguments{
\item{dat}{Prepared data frame}

\item{k_start}{Starting value for k}

\item{epsilon}{Epsilon value}

\item{stream_counts}{Named vector of stream counts}
}
\value{
Named list of starting values
}
\description{
Compute Starting Values for Saturated Joint Hurdle Model
}
\keyword{internal}
