% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{conssum}
\alias{conssum}
\title{Purchase Task Descriptive Summary}
\usage{
conssum(mat, x = x)
}
\arguments{
\item{mat}{A matrix of purchase task data.}

\item{x}{A vector of prices coinciding with purchase task data.}
}
\value{
A matrix of descriptive statistics.
}
\description{
Calculates descriptive statistics from purchase task data.
}
\details{
Provides the following descriptive statistics from purchase task data at each price: mean consumption, standard deviation of consumption, and percentage of 0 values.
}
\author{
Brent Kaplan <bkaplan4@ku.edu>
}

