% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{doEverything}
\alias{doEverything}
\title{Do Everything}
\usage{
doEverything(mat, prices, include0 = TRUE, equation, optimizer, k,
  remq0e = TRUE, replfree = NULL)
}
\arguments{
\item{mat}{Matrix of purchase task data.}

\item{prices}{Vector of prices coinciding with purchase task data}

\item{include0}{If FALSE, removes all 0s in consumption data prior to analysis. Default value is TRUE.}

\item{equation}{Character vector of length one. Accepts either "hs" for Hursh and Silberberg (2008) or "koff" for Koffarnus, Franck, Stein, and Bickel (2015). If "hs" and the first price (x) is 0, it will be replaced by replfree.}

\item{optimizer}{Character vector of length one. "default" will use the nonlinear (weighted) least-squares estimate and Gauss-Newton algorithm (for more information, see ?nls). "lm" will fit using the Levenberg-Marquardt algorithm (for more information, see ?nlsLM in the minpack.lm package). "brute" will fit using a two-stage process. First, a grid of potential starting values between q0 and alpha is created; 1 value of q0 being the maximum consumption value in the given dataframe and 100 values of alpha ranging from 0.00000001 to 0.1. Since q0 is recycled, the residual sum of squares is evaluated in each combination of q0 and alpha. The combination of of q0 and alpha that result in the least residual sum of squares is subsequently used as starting values in the nonlinear least-squares fitting process. The port algorithm is used to allow for lower and upper bounds. Lower and upper bounds for q0 are 0.001 and 1.5 times the maximum consumption; lower and upper bounds for alpha are 0.0000000001 and 1. For all fitting methods, a maximum of 1000 iterations are used.}

\item{k}{A numeric vector of length one. Reflects the range of consumption in log10 units.}

\item{remq0e}{If FALSE, retains consumption and price where price == 0. Default value is TRUE}

\item{replfree}{Optionally replaces price == 0 with specified value. Note, if fitting using equation == "hs", and 0 is first price, 0 gets replaced by replfree. Default value is .01}
}
\value{
List of results. Each element in the list contains the following subelements: [1] a dataframe used for fitting, [2] a model object, [3] a summary of the model object, [4] a dataframe of indices associated with the dataframe and model to be used in further demand analyses.
}
\description{
Analyzes matrix of purchase task data.
}
\details{
Analyzed matrix of purchase task data.
}
\author{
Brent Kaplan <bkaplan4@ku.edu>
}

