% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-registry.R
\docType{data}
\name{.beezdemand_equation_registry}
\alias{.beezdemand_equation_registry}
\title{Equation Registry}
\format{
An object of class \code{list} of length 12.
}
\usage{
.beezdemand_equation_registry
}
\description{
Maps equation IDs to their parameter requirements and scales.
}
\keyword{internal}
