% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-registry.R
\docType{data}
\name{.beezdemand_param_registry}
\alias{.beezdemand_param_registry}
\title{Canonical Parameter Definitions}
\format{
An object of class \code{list} of length 10.
}
\usage{
.beezdemand_param_registry
}
\description{
A list defining all canonical parameters, their constraints, and scale info.
}
\keyword{internal}
