% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmax-omax-engine.R
\name{.calc_observed_pmax_omax}
\alias{.calc_observed_pmax_omax}
\title{Calculate Observed Pmax and Omax for a Single Subject}
\usage{
.calc_observed_pmax_omax(price, consumption)
}
\arguments{
\item{price}{Numeric vector of prices}

\item{consumption}{Numeric vector of consumption values}
}
\value{
List with observed metrics
}
\description{
Calculate Observed Pmax and Omax for a Single Subject
}
\keyword{internal}
