% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{.tmb_anova_multifit}
\alias{.tmb_anova_multifit}
\title{Sequential LRT / AIC table for nested beezdemand_tmb fits}
\usage{
.tmb_anova_multifit(object, extra, test)
}
\description{
Sequential LRT / AIC table for nested beezdemand_tmb fits
}
\keyword{internal}
