% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{.tmb_group_terms}
\alias{.tmb_group_terms}
\title{Group beezdemand_tmb fixed effects into testable blocks for anova()}
\usage{
.tmb_group_terms(object, terms = NULL, group_by = "auto")
}
\description{
Group beezdemand_tmb fixed effects into testable blocks for anova()
}
\keyword{internal}
