% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{.tmb_predict_equation}
\alias{.tmb_predict_equation}
\title{Predict Single Observation for Each Equation}
\usage{
.tmb_predict_equation(price, Q0, alpha, k, log_q0, equation)
}
\description{
Predict Single Observation for Each Equation
}
\keyword{internal}
