% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{.tmb_term_assign_map}
\alias{.tmb_term_assign_map}
\title{Map a TMB design-matrix column to its originating model term}
\usage{
.tmb_term_assign_map(object, param)
}
\description{
Map a TMB design-matrix column to its originating model term
}
\keyword{internal}
