% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etm}
\alias{etm}
\title{Example Experimental Tobacco Marketplace data}
\format{
Long-form data.frame with columns: id, x, y, target, group. Participants were asked how many cigarettes, e-cigarettes, combustible, and non-combustible products they would buy at various prices.
}
\usage{
etm
}
\description{
A dataset containing ETM data for a small number of participants
}
\keyword{datasets}
