% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{Extractor}
\alias{Extractor}
\title{Extractor}
\usage{
Extractor(pid, adf, fit, eq, cols, kest)
}
\arguments{
\item{pid}{Participant ID}

\item{adf}{A data frame}

\item{fit}{Fitted model object}

\item{eq}{Equation specified}

\item{cols}{Column names to populate the dataframe row}

\item{kest}{Specification of k value}
}
\value{
One row of a dataframe with results
}
\description{
Extracts the useful information
}
\details{
Populates a single row of a dataframe consisting of important information from fits, etc.
}
\author{
Brent Kaplan <bkaplan.ku@gmail.com>
}
