% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint-hurdle.R
\name{fit_joint_latent}
\alias{fit_joint_latent}
\title{Fit Latent-Trait Joint Hurdle Model (Internal)}
\usage{
fit_joint_latent(
  dat,
  prep,
  n_subjects,
  stream_counts,
  k,
  k_fixed,
  start,
  epsilon,
  control,
  verbose,
  call
)
}
\description{
Fit Latent-Trait Joint Hurdle Model (Internal)
}
\keyword{internal}
