% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint-hurdle.R
\name{fit_joint_saturated}
\alias{fit_joint_saturated}
\title{Fit Saturated Joint Hurdle Model (Internal)}
\usage{
fit_joint_saturated(
  dat,
  prep,
  n_subjects,
  stream_counts,
  k_start,
  k_fixed,
  start,
  epsilon,
  control,
  verbose,
  call
)
}
\description{
Fit Saturated Joint Hurdle Model (Internal)
}
\keyword{internal}
