% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{fixef.beezdemand_tmb}
\alias{fixef.beezdemand_tmb}
\title{Extract Fixed Effects from TMB Model}
\usage{
\method{fixef}{beezdemand_tmb}(object, ...)
}
\arguments{
\item{object}{A \code{beezdemand_tmb} object.}

\item{...}{Additional arguments.}
}
\value{
Named numeric vector of fixed effects.
}
\description{
Extract Fixed Effects from TMB Model
}
\examples{
\donttest{
data(apt)
fit <- fit_demand_tmb(apt, equation = "exponential", verbose = 0)
nlme::fixef(fit)
}

}
