% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-methods.R
\name{fixef.cp_model_lmer}
\alias{fixef.cp_model_lmer}
\title{Extract Fixed Effects from Mixed-Effects Cross-Price Model}
\usage{
\method{fixef}{cp_model_lmer}(object, ...)
}
\arguments{
\item{object}{A cp_model_lmer object}

\item{...}{Additional arguments passed to fixef}
}
\value{
Named vector of fixed effects
}
\description{
Extract Fixed Effects from Mixed-Effects Cross-Price Model
}
