% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{GetAnalyticPmax}
\alias{GetAnalyticPmax}
\title{Get pmax}
\usage{
GetAnalyticPmax(Alpha, K, Q0)
}
\arguments{
\item{Alpha}{alpha parameter}

\item{K}{k parameter ( > lower limit )}

\item{Q0}{Q0}
}
\value{
Numeric
}
\description{
...
}
\details{
...
}
\examples{
GetAnalyticPmax(Alpha = 0.001, K = 3, Q0 = 10)
}
\author{
Shawn Gilroy \href{mailto:sgilroy1@lsu.edu}{sgilroy1@lsu.edu}
}
