% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{GetAnalyticPmaxFallback}
\alias{GetAnalyticPmaxFallback}
\title{Analytic Pmax Fallback}
\usage{
GetAnalyticPmaxFallback(K_, A_, Q0_)
}
\arguments{
\item{K_}{k parameter}

\item{A_}{alpha parameter}

\item{Q0_}{q0 parameter}
}
\value{
numeric
}
\description{
Fallback method for Analytic Pmax
}
\details{
Derivative-based optimization strategy
}
\examples{
GetAnalyticPmaxFallback(K_ = 1, A_ = 0.001, Q0_ = 10)
}
\author{
Shawn Gilroy \href{mailto:sgilroy1@lsu.edu}{sgilroy1@lsu.edu}
}
