% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{GetEmpirical}
\alias{GetEmpirical}
\title{GetEmpirical}
\usage{
GetEmpirical(dat, xcol = "x", ycol = "y", idcol = "id")
}
\arguments{
\item{dat}{data frame (long form) of purchase task data.}

\item{xcol}{The column name that should be treated as "x" data}

\item{ycol}{The column name that should be treated as "y" data}

\item{idcol}{The column name that should be treated as dataset identifier}
}
\value{
Data frame of empirical measures
}
\description{
Calculates empirical measures for purchase task data
}
\details{
`r lifecycle::badge("superseded")`

`GetEmpirical()` has been superseded by [get_empirical_measures()], which
provides a modern S3 interface with standardized methods (`print()`, `summary()`, `plot()`).
`GetEmpirical()` will continue to work but is no longer recommended for new code.

Will calculate and return the following empirical measures: Intensity, BP0, BP1, Omax, and Pmax
}
\examples{
## Obtain empirical measures
GetEmpirical(apt)
}
\seealso{
[get_empirical_measures()] for the modern interface
}
\author{
Brent Kaplan <bkaplan.ku@gmail.com>
}
