% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{GetK}
\alias{GetK}
\title{Get K}
\usage{
GetK(dat, mnrange = TRUE)
}
\arguments{
\item{dat}{Dataframe (long form)}

\item{mnrange}{Boolean for whether k should be calculated based on the mean range + .5}
}
\value{
Numeric
}
\description{
Calculates a k value by looking for the max/min consumption across entire dataset and adds .5 to that range
}
\details{
`r lifecycle::badge("superseded")`

`GetK()` has been superseded by [get_k()], which provides explicit parameters
for the adjustment value and optional verbose output for better transparency.
`GetK()` will continue to work but is no longer recommended for new code.

Will look for maximum/minimum greater zero
}
\examples{
GetK(apt)
}
\seealso{
[get_k()] for the modern interface
}
\author{
Brent Kaplan <bkaplan.ku@gmail.com>
}
