% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{GetSharedK}
\alias{GetSharedK}
\title{Get Shared K}
\usage{
GetSharedK(dat, equation)
}
\arguments{
\item{dat}{Dataframe (longform)}

\item{equation}{Character vector. Accepts either "hs" or "koff"}
}
\value{
Numeric value of shared k
}
\description{
Finds shared k among selected datasets using global regression
}
\details{
Uses global regression to fit a shared k among datasets. Assumes the dataset is in its final form. Used within FitCurves
}
\author{
Brent Kaplan <bkaplan.ku@gmail.com>
}
