% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-hurdle-methods.R
\name{glance.beezdemand_cp_hurdle}
\alias{glance.beezdemand_cp_hurdle}
\title{Glance method for beezdemand_cp_hurdle}
\usage{
\method{glance}{beezdemand_cp_hurdle}(x, ...)
}
\arguments{
\item{x}{A beezdemand_cp_hurdle object}

\item{...}{Additional arguments (ignored)}
}
\value{
A one-row data frame of model statistics
}
\description{
Glance method for beezdemand_cp_hurdle
}
