% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-demand.R
\name{glance.beezdemand_fixed}
\alias{glance.beezdemand_fixed}
\title{Glance Method for beezdemand_fixed}
\usage{
\method{glance}{beezdemand_fixed}(x, ...)
}
\arguments{
\item{x}{A beezdemand_fixed object}

\item{...}{Additional arguments (ignored)}
}
\value{
A one-row tibble of model statistics
}
\description{
Glance Method for beezdemand_fixed
}
