% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint-hurdle-methods.R
\name{glance.beezdemand_joint_hurdle}
\alias{glance.beezdemand_joint_hurdle}
\title{Glance Method for Joint Hurdle Model}
\usage{
\method{glance}{beezdemand_joint_hurdle}(x, ...)
}
\arguments{
\item{x}{A \code{beezdemand_joint_hurdle} object.}

\item{...}{Additional arguments (unused).}
}
\value{
A tibble with model-level statistics.
}
\description{
Glance Method for Joint Hurdle Model
}
