% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic-wrappers.R
\name{glance.beezdemand_systematicity}
\alias{glance.beezdemand_systematicity}
\title{Glance Method for beezdemand_systematicity}
\usage{
\method{glance}{beezdemand_systematicity}(x, ...)
}
\arguments{
\item{x}{A beezdemand_systematicity object}

\item{...}{Additional arguments (ignored)}
}
\value{
A one-row tibble of overall statistics
}
\description{
Glance Method for beezdemand_systematicity
}
