% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-methods.R
\name{glance.cp_model_lm}
\alias{glance.cp_model_lm}
\title{Get model summaries from a linear cross-price model}
\usage{
\method{glance}{cp_model_lm}(x, ...)
}
\arguments{
\item{x}{A cp_model_lm object.}

\item{...}{Additional arguments (unused).}
}
\value{
A tibble with model summary statistics.
}
\description{
Get model summaries from a linear cross-price model
}
