% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-hurdle-utils.R
\name{interpret_cp_interaction}
\alias{interpret_cp_interaction}
\title{Interpret Cross-Price Interaction Parameter}
\usage{
interpret_cp_interaction(I_value)
}
\arguments{
\item{I_value}{The I parameter value (scalar or vector)}
}
\value{
Character vector with interpretation
}
\description{
Provides interpretation of the I parameter value.
}
