% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint-hurdle-methods.R
\name{joint-hurdle-methods}
\alias{joint-hurdle-methods}
\title{S3 Methods for Joint Hurdle Cross-Price Models}
\description{
Print, summary, predict, and plot methods for objects of class
  \code{beezdemand_joint_hurdle}.
}
