% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ko}
\alias{ko}
\title{Example nonhuman demand data with drug and dose}
\format{
A tibble with 135 rows and 6 columns:
\describe{
\item{monkey}{unique identifier}
\item{x}{fixed-ratio requirement}
\item{y}{consumption}
\item{y_ll4}{consumption transformed via ll4}
\item{drug}{drug}
\item{dose}{dose}
}
}
\usage{
ko
}
\description{
Data from Ko et al. (2002)
}
\keyword{datasets}
