% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-hurdle-methods.R
\name{logLik.beezdemand_cp_hurdle}
\alias{logLik.beezdemand_cp_hurdle}
\title{Log-likelihood method for beezdemand_cp_hurdle}
\usage{
\method{logLik}{beezdemand_cp_hurdle}(object, ...)
}
\arguments{
\item{object}{A beezdemand_cp_hurdle object}

\item{...}{Additional arguments (ignored)}
}
\description{
Log-likelihood method for beezdemand_cp_hurdle
}
