% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lowNicClean}
\alias{lowNicClean}
\title{Low-nicotine cigarette purchase task}
\format{
A data frame with 5 columns: id, condition, x, y, commodity.
}
\usage{
lowNicClean
}
\description{
Long-form purchase task data across nicotine conditions.
}
\examples{
# data(lowNicClean)
}
\keyword{datasets}
