% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lseq}
\alias{lseq}
\title{Create Logarithmically Spaced Sequence}
\usage{
lseq(from = 1e-10, to = 1000, length.out = 14)
}
\arguments{
\item{from}{Starting value. Default value is 0.0000000001.}

\item{to}{Ending value. Default value is 1000.}

\item{length.out}{Number of values in vector. Default value is 14.}
}
\value{
Vector
}
\description{
Creates a logarithmically spaced sequence.
}
\details{
A vector of logarithmically spaced values primarily used for plotting. Taken from library("emdbook").
}
\author{
library("emdbook")
}

