% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{makeSumTable}
\alias{makeSumTable}
\title{Make Summary Table}
\usage{
makeSumTable(reslst)
}
\arguments{
\item{reslst}{The return from doEverything.}
}
\value{
A dataframe with the following indices: empirical q0 (q0e), derived q0 (q0d), alpha (alpha), essential value (ev), empirical pmax (pmaxe), derived pmax using empirical q0 (pmaxdq0e), derived pmax using derived q0 (pmaxdq0d), empirical omax (omaxe), derived omax using empirical q0e (omaxdq0e), derived omax using derived q0 (omaxdq0d), auc using maximum consumption from entire sample as normalization (aucallmax), auc using individual maximum consumption as normalization (aucindmax), k (k), standard error of derived q0 (q0se), standard error of alpha (alphase), Sy.x (Sy.x), number of points analyzed (pointsanalyzed), degrees of freedom (df), convergence notes from model (convergnotes), equation used in fitting (eq), optimizer used in fitting (nltype), whether empirical q0 was used in fitting (remq0e), whether a parallel shift was used (parshift), replacement number for 0s if applicable (replnum), whether 0s were included in fitting (incl0s).
}
\description{
Makes a summary table.
}
\details{
Makes a summary table from the return from doEverything.
}
\author{
Brent Kaplan <bkaplan4@ku.edu>
}

