% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{minTicks}
\alias{minTicks}
\title{Create Minor Tick Sequence}
\usage{
minTicks(maj)
}
\arguments{
\item{maj}{Value from function lseq}
}
\value{
Vector
}
\description{
Creates minor ticks for use in plotting.
}
\author{
Brent Kaplan <bkaplan.ku@gmail.com>
}
