% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic-wrappers.R
\name{new_beezdemand_systematicity}
\alias{new_beezdemand_systematicity}
\title{Create a beezdemand_systematicity Object}
\usage{
new_beezdemand_systematicity(results, type, call, by_var = NULL)
}
\arguments{
\item{results}{Tibble with standardized columns}

\item{type}{Character: "demand" or "cp"}

\item{call}{Original function call}

\item{by_var}{Character vector of grouping column names (NULL if ungrouped)}
}
\description{
Create a beezdemand_systematicity Object
}
\keyword{internal}
