% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ongoingETM}
\alias{ongoingETM}
\title{Experimental Tobacco Marketplace (ETM) data}
\format{
A data frame with 6 columns: id, x, AdjCig, FixCig, ECig, flavor.
}
\usage{
ongoingETM
}
\description{
ETM data across price points with product quantities.
}
\examples{
# data(ongoingETM)
}
\keyword{datasets}
