% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint-hurdle-methods.R
\name{plot.beezdemand_joint_hurdle}
\alias{plot.beezdemand_joint_hurdle}
\title{Plot Joint Hurdle Model}
\usage{
\method{plot}{beezdemand_joint_hurdle}(
  x,
  type = c("demand", "probability", "individual", "parameters"),
  stream = c("all", "alone.target", "own.target", "own.alt"),
  ...
)
}
\arguments{
\item{x}{A \code{beezdemand_joint_hurdle} object.}

\item{type}{One of "demand", "probability", "individual", "parameters".}

\item{stream}{Which stream(s) to plot.}

\item{...}{Additional arguments passed to ggplot.}
}
\value{
A ggplot2 object.
}
\description{
Plot Joint Hurdle Model
}
