% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCurve}
\alias{PlotCurve}
\title{Plot Curve}
\usage{
PlotCurve(adf, dfrow, fitfail)
}
\arguments{
\item{adf}{Data frame (long form) of purchase task data.}

\item{dfrow}{A row of results from FitCurves}

\item{fitfail}{Boolean whether there's a valid nls model object}
}
\value{
ggplot2 graphical object
}
\description{
Creates a single plot object
}
\details{
Creates individual demand curves
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu>
}
