% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCurves}
\alias{PlotCurves}
\title{Plot Curves}
\usage{
PlotCurves(adf, dfrow, fit, outdir = "../plots/", fitfail, tobquote, vartext)
}
\arguments{
\item{adf}{Data frame (long form) of purchase task data.}

\item{dfrow}{A row of results from FitCurves}

\item{fit}{A nls model object from FitCurves}

\item{outdir}{Directory where plots are saved}

\item{fitfail}{Boolean whether there's a valid nls model object}

\item{tobquote}{Character string to be evaluated}

\item{vartext}{Character vector to match demand indices}
}
\value{
Nothing
}
\description{
Creates plots
}
\details{
Creates and saves plots of individual demand curves
}
\author{
Brent Kaplan <bkaplan4@ku.edu>
}

