% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotInd}
\alias{plotInd}
\title{Plot Individual Curves}
\usage{
plotInd(reslst, basedir = "../", basename = "indplots/", work = FALSE,
  vartext = NULL)
}
\arguments{
\item{reslst}{Return from doEverything}

\item{basedir}{Base directory}

\item{basename}{Directory to have plots saved in}

\item{work}{If TRUE, work function will be plotted. Default value is FALSE.}

\item{vartext}{Character vector of variables to be reported on plots.}
}
\value{
Message indicating how many plots were made and where plots were saved.
}
\description{
Plots individual demand curves.
}
\details{
Plots and saves individual demand curves.
}
\author{
Brent Kaplan <bkaplan4@ku.edu>
}

