% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-methods.R
\name{predict.cp_model_lm}
\alias{predict.cp_model_lm}
\title{Predict method for cp_model_lm objects.}
\usage{
\method{predict}{cp_model_lm}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A cp_model_lm object.}

\item{newdata}{Data frame containing new x values.}

\item{...}{Additional arguments.}
}
\value{
Data frame with predictions.
}
\description{
Predict method for cp_model_lm objects.
}
