% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-methods.R
\name{predict.cp_model_nls}
\alias{predict.cp_model_nls}
\title{Predict from a Cross-Price Demand Model (Nonlinear)}
\usage{
\method{predict}{cp_model_nls}(
  object,
  newdata = NULL,
  inv_fun = identity,
  inverse_fun = deprecated(),
  ...
)
}
\arguments{
\item{object}{A cross-price model object from fit_cp_nls with return_all=TRUE.}

\item{newdata}{A data frame containing an 'x' column.}

\item{inv_fun}{Optional inverse transformation function. Default is \code{identity}.}

\item{inverse_fun}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{inv_fun} instead.}

\item{...}{Additional arguments.}
}
\value{
A data frame with x values and predicted y values.
}
\description{
Predict from a Cross-Price Demand Model (Nonlinear)
}
