% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-comparison.R
\name{print.anova.beezdemand_hurdle}
\alias{print.anova.beezdemand_hurdle}
\title{Print Method for ANOVA Comparisons}
\usage{
\method{print}{anova.beezdemand_hurdle}(x, digits = 4, ...)
}
\arguments{
\item{x}{An `anova.beezdemand` object.}

\item{digits}{Number of significant digits to print.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for ANOVA Comparisons
}
