% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.beezdemand}
\alias{print.beezdemand}
\title{print.beezdemand}
\usage{
\method{print}{beezdemand}(dfres)
}
\arguments{
\item{dfres}{resulting data frame}
}
\value{

}
\description{
On print, show data frame, otherwise beez obj
}
