% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-hurdle-methods.R
\name{print.beezdemand_cp_hurdle}
\alias{print.beezdemand_cp_hurdle}
\title{Print method for beezdemand_cp_hurdle}
\usage{
\method{print}{beezdemand_cp_hurdle}(x, ...)
}
\arguments{
\item{x}{A beezdemand_cp_hurdle object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for beezdemand_cp_hurdle
}
