% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.beezdemand_diagnostic_plots}
\alias{print.beezdemand_diagnostic_plots}
\title{Print Method for Diagnostic Plots}
\usage{
\method{print}{beezdemand_diagnostic_plots}(x, ...)
}
\arguments{
\item{x}{A \code{beezdemand_diagnostic_plots} object (list of ggplot objects).}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for Diagnostic Plots
}
