% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.beezdemand_diagnostics}
\alias{print.beezdemand_diagnostics}
\title{Print Method for Model Diagnostics}
\usage{
\method{print}{beezdemand_diagnostics}(x, ...)
}
\arguments{
\item{x}{A `beezdemand_diagnostics` object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for Model Diagnostics
}
