% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-demand.R
\name{print.beezdemand_fixed}
\alias{print.beezdemand_fixed}
\title{Print Method for beezdemand_fixed}
\usage{
\method{print}{beezdemand_fixed}(x, ...)
}
\arguments{
\item{x}{A beezdemand_fixed object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for beezdemand_fixed
}
