% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-methods.R
\name{print.beezdemand_fixed_grouped}
\alias{print.beezdemand_fixed_grouped}
\title{Print Method for beezdemand_fixed_grouped}
\usage{
\method{print}{beezdemand_fixed_grouped}(x, ...)
}
\arguments{
\item{x}{A beezdemand_fixed_grouped object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for beezdemand_fixed_grouped
}
