% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdle-methods.R
\name{print.beezdemand_hurdle}
\alias{print.beezdemand_hurdle}
\title{Print Method for Hurdle Demand Model}
\usage{
\method{print}{beezdemand_hurdle}(x, ...)
}
\arguments{
\item{x}{An object of class \code{beezdemand_hurdle}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print Method for Hurdle Demand Model
}
