% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-comparison.R
\name{print.beezdemand_model_comparison}
\alias{print.beezdemand_model_comparison}
\title{Print Method for Model Comparison}
\usage{
\method{print}{beezdemand_model_comparison}(x, digits = 4, ...)
}
\arguments{
\item{x}{A `beezdemand_model_comparison` object.}

\item{digits}{Number of significant digits to print.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for Model Comparison
}
