% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic-wrappers.R
\name{print.beezdemand_systematicity}
\alias{print.beezdemand_systematicity}
\title{Print Method for beezdemand_systematicity}
\usage{
\method{print}{beezdemand_systematicity}(x, ...)
}
\arguments{
\item{x}{A beezdemand_systematicity object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for beezdemand_systematicity
}
