% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-hurdle-utils.R
\name{print.cp_hurdle_comparison}
\alias{print.cp_hurdle_comparison}
\title{Print method for cp_hurdle_comparison}
\usage{
\method{print}{cp_hurdle_comparison}(x, ...)
}
\arguments{
\item{x}{A cp_hurdle_comparison object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for cp_hurdle_comparison
}
