% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-methods.R
\name{print.cp_model_lm}
\alias{print.cp_model_lm}
\title{Print a Cross-Price Demand Model (Linear)}
\usage{
\method{print}{cp_model_lm}(x, ...)
}
\arguments{
\item{x}{A \code{cp_model_lm} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a Cross-Price Demand Model (Linear)
}
