% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-methods.R
\name{print.cp_model_lmer}
\alias{print.cp_model_lmer}
\title{Print a Cross-Price Demand Model (Mixed-Effects)}
\usage{
\method{print}{cp_model_lmer}(x, ...)
}
\arguments{
\item{x}{A \code{cp_model_lmer} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a Cross-Price Demand Model (Mixed-Effects)
}
