% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-methods.R
\name{print.cp_model_nls}
\alias{print.cp_model_nls}
\title{Print a Cross-Price Demand Model (Nonlinear)}
\usage{
\method{print}{cp_model_nls}(x, ...)
}
\arguments{
\item{x}{A \code{cp_model_nls} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a Cross-Price Demand Model (Nonlinear)
}
