% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-methods.R
\name{print.cp_posthoc}
\alias{print.cp_posthoc}
\title{Print method for cp_posthoc objects}
\usage{
\method{print}{cp_posthoc}(x, ...)
}
\arguments{
\item{x}{A cp_posthoc object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for cp_posthoc objects
}
