% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-demand.R
\name{print.summary.beezdemand_fixed}
\alias{print.summary.beezdemand_fixed}
\title{Print Method for summary.beezdemand_fixed}
\usage{
\method{print}{summary.beezdemand_fixed}(x, digits = 4, ...)
}
\arguments{
\item{x}{A summary.beezdemand_fixed object}

\item{digits}{Number of significant digits to print}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for summary.beezdemand_fixed
}
