% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdle-methods.R
\name{print.summary.beezdemand_hurdle}
\alias{print.summary.beezdemand_hurdle}
\title{Print Summary of Hurdle Demand Model}
\usage{
\method{print}{summary.beezdemand_hurdle}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{summary.beezdemand_hurdle}.}

\item{digits}{Number of significant digits to print. Default is 4.}

\item{...}{Additional arguments passed to \code{print}.}
}
\description{
Print Summary of Hurdle Demand Model
}
