% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic-wrappers.R
\name{print.summary.beezdemand_systematicity}
\alias{print.summary.beezdemand_systematicity}
\title{Print Method for summary.beezdemand_systematicity}
\usage{
\method{print}{summary.beezdemand_systematicity}(x, n = 20, ...)
}
\arguments{
\item{x}{A summary.beezdemand_systematicity object}

\item{n}{Number of IDs to display (default 20)}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for summary.beezdemand_systematicity
}
