% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-methods.R
\name{print.summary.cp_model_lm}
\alias{print.summary.cp_model_lm}
\title{Print method for summary.cp_model_lm objects.}
\usage{
\method{print}{summary.cp_model_lm}(x, ...)
}
\arguments{
\item{x}{A `summary.*` object}

\item{...}{Unused}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for summary.cp_model_lm objects.
}
