% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-methods.R
\name{print.summary.cp_model_lmer}
\alias{print.summary.cp_model_lmer}
\title{Print method for summary.cp_model_lmer objects.}
\usage{
\method{print}{summary.cp_model_lmer}(x, ...)
}
\arguments{
\item{x}{A \verb{summary.*} object}

\item{...}{Unused}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print method for summary.cp_model_lmer objects.
}
