% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-methods.R
\name{print.summary.cp_model_nls}
\alias{print.summary.cp_model_nls}
\title{Print method for summary.cp_model_nls objects}
\usage{
\method{print}{summary.cp_model_nls}(x, ...)
}
\arguments{
\item{x}{A `summary.*` object}

\item{...}{Unused}
}
\description{
Print method for summary.cp_model_nls objects
}
