% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic-methods.R
\name{print.summary.cp_unsystematic}
\alias{print.summary.cp_unsystematic}
\title{Print Method for Cross-Price Unsystematic Summary}
\usage{
\method{print}{summary.cp_unsystematic}(x, ...)
}
\arguments{
\item{x}{A \code{summary.cp_unsystematic} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for summary.cp_unsystematic
}
