% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull}
\alias{pull}
\title{Pull}
\usage{
pull(x, y)
}
\arguments{
\item{x}{A data frame}

\item{y}{Name of column}
}
\value{
Vector
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Pulls a single vector from a data frame. Good to use with dplyr.
From https://stackoverflow.com/questions/21618423/extract-a-dplyr-tbl-column-as-a-vector
}
\details{
Pull vector from data frame
}
\author{
Brent Kaplan \href{mailto:bkaplan.ku@gmail.com}{bkaplan.ku@gmail.com}
}
