% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed-methods.R
\name{ranef.beezdemand_nlme}
\alias{ranef.beezdemand_nlme}
\title{Extract Random Effects from a beezdemand_nlme Model}
\usage{
\method{ranef}{beezdemand_nlme}(object, ...)
}
\arguments{
\item{object}{A \code{beezdemand_nlme} object.}

\item{...}{Additional arguments passed to \code{nlme::ranef()}.}
}
\value{
A data frame (or list of data frames if multiple levels of grouping)
of random effects, as returned by \code{ranef.nlme()}.
}
\description{
S3 method for \code{ranef} for objects of class \code{beezdemand_nlme}.
Extracts the random effects from the fitted \code{nlme} model.
}
\seealso{
\code{\link{coef.beezdemand_nlme}}, \code{\link{fixef.beezdemand_nlme}}
}
