% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-methods.R
\name{ranef.cp_model_lmer}
\alias{ranef.cp_model_lmer}
\title{Extract Random Effects from Mixed-Effects Cross-Price Model}
\usage{
\method{ranef}{cp_model_lmer}(object, ...)
}
\arguments{
\item{object}{A cp_model_lmer object}

\item{...}{Additional arguments passed to ranef}
}
\value{
List of random effects
}
\description{
Extract Random Effects from Mixed-Effects Cross-Price Model
}
