% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{RecodeOutliers}
\alias{RecodeOutliers}
\title{Recode Outliers}
\usage{
RecodeOutliers(x, outval = 3.29, unitshigher = 1)
}
\arguments{
\item{x}{A vector}

\item{outval}{Values greater than or equal to this number (specified in standard deviations) will be recoded. Default is 3.29SD as specified by Tabachnick and Fidell (2013)}

\item{unitshigher}{Outliers identified by outval will be coded to a certain number of units higher than the greatest nonoutlier value. Default is 1 unit as specified by Tabachnick and Fidell (2013)}
}
\value{
A vector with recoded values alongside original values
}
\description{
Recodes outliers
}
\details{
Recodes outliers using Tabachnick and Fidell's (2013) criteria. A variable is standardized and values that are greater than or equal to a specified outlier value (specified in standard deviations; default 3.29SD) are recoded to a certain number of units (default 1) higher than the greatest nonoutlier value.
}
\author{
Brent Kaplan <bkaplan4@ku.edu>
}

