% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{glance}
\alias{augment}
\alias{VarCorr}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom:reexports]{augment}}, \code{\link[broom:reexports]{glance}}, \code{\link[broom:reexports]{tidy}}}

  \item{lme4}{\code{\link[lme4]{VarCorr}}}
}}

