% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{ReplaceZeros}
\alias{ReplaceZeros}
\title{Replace Zeros}
\usage{
ReplaceZeros(dat, nrepl = 1, replnum = 0.01)
}
\arguments{
\item{dat}{Dataframe (long form)}

\item{nrepl}{Number of zeros to replace with replacement value (replnum). Can accept either a number or "all" if all zeros should be replaced. Default is to replace the first zero only.}

\item{replnum}{Value to replace zeros. Default is .01}
}
\value{
Dataframe (long form)
}
\description{
Replaces 0 values
}
\details{
Replaces specified number of 0s with replacement value.
}
\examples{
ReplaceZeros(apt)
## Replace all zeros with .01
ReplaceZeros(apt, nrepl = "all", replnum = .01)
}
\author{
Brent Kaplan <bkaplan4@ku.edu>
}

