% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-theme.R
\name{scale_color_beezdemand}
\alias{scale_color_beezdemand}
\title{beezdemand Color Scale (Discrete)}
\usage{
scale_color_beezdemand(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{ggplot2::scale_color_manual}.}
}
\value{
A ggplot2 discrete color scale.
}
\description{
beezdemand Color Scale (Discrete)
}
