% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-theme.R
\name{scale_fill_beezdemand}
\alias{scale_fill_beezdemand}
\title{beezdemand Fill Scale (Discrete)}
\usage{
scale_fill_beezdemand(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{ggplot2::scale_fill_manual}.}
}
\value{
A ggplot2 discrete fill scale.
}
\description{
beezdemand Fill Scale (Discrete)
}
\examples{
\donttest{
library(ggplot2)
ggplot(iris, aes(Species, Sepal.Length, fill = Species)) +
  geom_boxplot() +
  scale_fill_beezdemand()
}
}
