% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{splitMat}
\alias{splitMat}
\title{Split Matrix}
\usage{
splitMat(mat, x = x, incl0 = TRUE, replfree = NULL)
}
\arguments{
\item{mat}{A matrix in wide form where each column is a participant's responses. If a dataframe is provided, it will attempt to coerce it into a matrix.}

\item{x}{A vector of prices used in the purchase task. The number of elements must be equal to the number of rows in the matrix.}

\item{incl0}{If FALSE, ALL 0 values and their associated prices will be removed, even those that precede reverals from 0.}

\item{replfree}{Optionally replaces price=0 with specified number.}
}
\value{
A list of dataframes and/or character vectors whose length is equal to the number of columns (i.e., participants) in the original matrix. Prints the number of cases that contain no consumption and fewer than 3 datapoints.
}
\description{
Splits a matrix of purchase task data into individual dataframes consisting of price, consumption, and expenditure.
}
\details{
This function takes a matrix and splits it into a list of dataframes. Each dataframe consists of prices (x), consumption values (y), and expenditure values (expend). If a participant has no consumption, the dataframe will instead be a character vector indicating "No consumption." If a participant has fewer than 3 datapoints, the dataframe will instead be a character vector indicating "Fewer than 3 datapoints." Note that these conditions are more likely to be met when 0s are deleted.
}
\author{
Brent Kaplan <bkaplan4@ku.edu>
}

