% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic-wrappers.R
\name{summary.beezdemand_systematicity}
\alias{summary.beezdemand_systematicity}
\title{Summary Method for beezdemand_systematicity}
\usage{
\method{summary}{beezdemand_systematicity}(object, ...)
}
\arguments{
\item{object}{A beezdemand_systematicity object}

\item{...}{Additional arguments (ignored)}
}
\value{
A summary.beezdemand_systematicity object
}
\description{
Summary Method for beezdemand_systematicity
}
