% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-methods.R
\name{summary.cp_model_lm}
\alias{summary.cp_model_lm}
\title{Summary method for cp_model_lm objects.}
\usage{
\method{summary}{cp_model_lm}(object, ...)
}
\arguments{
\item{object}{A cp_model_lm object.}

\item{...}{Additional arguments.}
}
\value{
A list summarizing the linear model.
}
\description{
Summary method for cp_model_lm objects.
}
