% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-methods.R
\name{summary.cp_model_lmer}
\alias{summary.cp_model_lmer}
\title{Summary method for cp_model_lmer objects.}
\usage{
\method{summary}{cp_model_lmer}(object, ...)
}
\arguments{
\item{object}{A cp_model_lmer object.}

\item{...}{Additional arguments.}
}
\value{
A list summarizing the mixed-effects model.
}
\description{
Summary method for cp_model_lmer objects.
}
