% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{systemCheck}
\alias{systemCheck}
\title{Systematic Purchase Task Data Checker}
\usage{
systemCheck(mat = NULL, x = NULL, deltaq = 0.025, bounce = 0.1,
  reversals = 0, minpos = 4)
}
\arguments{
\item{mat}{A matrix in wide form where each column is a participant's responses. If a dataframe is provided, it will attempt to coerce it to a matrix.}

\item{x}{A vector of prices used in the purchase task. The number of elements must be equal to the number of rows in the matrix.}

\item{deltaq}{Numeric vector of length equal to one. The criterion by which the relative change in quantity purchased will be compared. Relative changes in quantity purchased below this criterion will be flagged. Default value is 0.025.}

\item{bounce}{Numeric vector of length equal to one. The criterion by which the number of price-to-price increases in consumption that exceed 25\% of initial consumption at the lowest price, expressed relative to the total number of price increments, will be compared. The relative number of price-to-price increases above this criterion will be flagged. Default value is 0.10.}

\item{reversals}{Numeric vector of length equal to one. The criterion by which the number of reversals from 0 will be compared. Number of reversals above this criterion will be flagged. Default value is 0.}

\item{minpos}{Minimum positive values to be checked against. Default value if 4.}
}
\value{
A dataframe providing individual values of deltaq, bounce, reversals, and number of positive consumption values. Also contains passing information for each criterion, as well as a total number of passes (total number of passes applies only to Stein et al.'s criteria). Prints the number of cases and the percentage of cases that fail at least one criterion.
}
\description{
Applies Stein, Koffarnus, Snider, Quisenberry, & Bickel's (2015) criteria for identification of nonsystematic purchase task data.
}
\details{
This function applies the 3 criteria proposed by Stein et al., (2015) for identification of nonsystematic purchase task data. The three criteria include trend (deltaq), bounce, and reversals from 0. Also flags for a minimum number of positive consumption values.
}
\author{
Brent Kaplan <bkaplan4@ku.edu>
}

