% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-theme.R
\name{theme_beezdemand}
\alias{theme_beezdemand}
\title{beezdemand Plot Theme}
\usage{
theme_beezdemand(
  style = c("modern", "apa"),
  base_size = 11,
  base_family = "sans"
)
}
\arguments{
\item{style}{Character. One of \code{"modern"} or \code{"apa"}.}

\item{base_size}{Base font size (default: 11).}

\item{base_family}{Base font family (default: "sans").}
}
\value{
A ggplot2 theme object.
}
\description{
beezdemand Plot Theme
}
\examples{
\donttest{
library(ggplot2)
ggplot(iris, aes(Sepal.Length, Sepal.Width)) +
  geom_point() +
  theme_beezdemand()
}
}
