% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint-hurdle-methods.R
\name{tidy.beezdemand_joint_hurdle}
\alias{tidy.beezdemand_joint_hurdle}
\title{Tidy Method for Joint Hurdle Model}
\usage{
\method{tidy}{beezdemand_joint_hurdle}(x, ...)
}
\arguments{
\item{x}{A \code{beezdemand_joint_hurdle} object.}

\item{...}{Additional arguments (unused).}
}
\value{
A tibble with coefficient information.
}
\description{
Tidy Method for Joint Hurdle Model
}
