% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic-wrappers.R
\name{tidy.beezdemand_systematicity}
\alias{tidy.beezdemand_systematicity}
\title{Tidy Method for beezdemand_systematicity}
\usage{
\method{tidy}{beezdemand_systematicity}(x, ...)
}
\arguments{
\item{x}{A beezdemand_systematicity object}

\item{...}{Additional arguments (ignored)}
}
\value{
The per-subject results tibble
}
\description{
Tidy Method for beezdemand_systematicity
}
