% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim.leading}
\alias{trim.leading}
\title{Trim Leading Characters}
\usage{
trim.leading(x)
}
\arguments{
\item{x}{A string}
}
\value{
Character vector of length 1
}
\description{
Trims leading characters
}
\details{
Trims leading characters
}
\author{
Brent Kaplan <bkaplan.ku@gmail.com>
}
