% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beezdemand-package.R
\docType{package}
\name{beezdemand-package}
\alias{beezdemand}
\alias{beezdemand-package}
\title{beezdemand: Behavioral Economic Easy Demand}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Facilitates many of the analyses performed in studies of behavioral economic demand. The package supports commonly-used options for modeling operant demand including (1) data screening proposed by Stein, Koffarnus, Snider, Quisenberry, & Bickel (2015; \doi{10.1037/pha0000020}), (2) fitting models of demand such as linear (Hursh, Raslear, Bauman, & Black, 1989, \doi{10.1007/978-94-009-2470-3_22}), exponential (Hursh & Silberberg, 2008, \doi{10.1037/0033-295X.115.1.186}) and modified exponential (Koffarnus, Franck, Stein, & Bickel, 2015, \doi{10.1037/pha0000045}), and (3) calculating numerous measures relevant to applied behavioral economists (Intensity, Pmax, Omax). Also supports plotting and comparing data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://brentkaplan.github.io/beezdemand/}
  \item \url{https://github.com/brentkaplan/beezdemand}
  \item Report bugs at \url{https://github.com/brentkaplan/beezdemand/issues}
}

}
\author{
\strong{Maintainer}: Brent Kaplan \email{bkaplan.ku@gmail.com} (\href{https://orcid.org/0000-0002-3758-6776}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Shawn Gilroy \email{shawn.gilroy@temple.edu} [contributor]
}

}
\keyword{internal}
