% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{calc_group_metrics.beezdemand_tmb}
\alias{calc_group_metrics.beezdemand_tmb}
\title{Calculate Population-Level Demand Metrics for TMB Model}
\usage{
\method{calc_group_metrics}{beezdemand_tmb}(object, ...)
}
\arguments{
\item{object}{A \code{beezdemand_tmb} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A list with Pmax, Omax, Qmax, elasticity_at_pmax, and method.
}
\description{
Calculate Population-Level Demand Metrics for TMB Model
}
\examples{
\donttest{
data(apt)
fit <- fit_demand_tmb(apt, equation = "exponential", verbose = 0)
calc_group_metrics(fit)
}

}
