% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckCols}
\alias{CheckCols}
\title{Check Column Names}
\usage{
CheckCols(dat, xcol, ycol, idcol, groupcol = NULL)
}
\arguments{
\item{dat}{Dataframe}

\item{xcol}{Name of x column}

\item{ycol}{Name of y column}

\item{idcol}{Name of id column}

\item{groupcol}{Name of group column}
}
\value{
Dataframe
}
\description{
Checks to ensure column names are specified
}
\details{
Check column names
}
\author{
Brent Kaplan <bkaplan.ku@gmail.com>
}
