% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{CheckUnsystematic}
\alias{CheckUnsystematic}
\title{Systematic Purchase Task Data Checker}
\usage{
CheckUnsystematic(dat, deltaq = 0.025, bounce = 0.1, reversals = 0,
  ncons0 = 2)
}
\arguments{
\item{dat}{Dataframe in long form. Colums are id, x, y.}

\item{deltaq}{Numeric vector of length equal to one. The criterion by which the relative change in quantity purchased will be compared. Relative changes in quantity purchased below this criterion will be flagged. Default value is 0.025.}

\item{bounce}{Numeric vector of length equal to one. The criterion by which the number of price-to-price increases in consumption that exceed 25\% of initial consumption at the lowest price, expressed relative to the total number of price increments, will be compared. The relative number of price-to-price increases above this criterion will be flagged. Default value is 0.10.}

\item{reversals}{Numeric vector of length equal to one. The criterion by which the number of reversals from number of consecutive (see ncons0) 0s will be compared. Number of reversals above this criterion will be flagged. Default value is 0.}

\item{ncons0}{Numer of consecutive 0s prior to a positive value is used to flag for a reversal. Value can be either 1 (relatively more conservative) or 2 (default; as recommended by Stein et al., (2015).}
}
\value{
Dataframe
}
\description{
Applies Stein, Koffarnus, Snider, Quisenberry, & Bickel's (2015) criteria for identification of nonsystematic purchase task data.
}
\details{
This function applies the 3 criteria proposed by Stein et al., (2015) for identification of nonsystematic purchase task data. The three criteria include trend (deltaq), bounce, and reversals from 0. Also reports number of positive consumption values.
}
\author{
Brent Kaplan <bkaplan4@ku.edu>
}

