% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{coef.beezdemand_tmb}
\alias{coef.beezdemand_tmb}
\title{Extract Coefficients from TMB Model}
\usage{
\method{coef}{beezdemand_tmb}(object, ...)
}
\arguments{
\item{object}{A \code{beezdemand_tmb} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Named numeric vector of fixed effect coefficients.
}
\description{
Extract Coefficients from TMB Model
}
\examples{
\donttest{
data(apt)
fit <- fit_demand_tmb(apt, equation = "exponential", verbose = 0)
coef(fit)
}

}
