% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cp}
\alias{cp}
\title{Example cross‐price dataset}
\format{
A data frame with N rows and columns:
  \describe{
    \item{id}{unique identifier}
    \item{x}{price of the alternative}
    \item{y}{consumption/demand}
    \item{target}{target (e.g., alone, own, alt)}
    \item{group}{e.g., drug or product}
  }
}
\usage{
cp
}
\description{
A small illustrative dataset of price (x) and consumption (y) target
(target), and group (group).
}
\keyword{datasets}
