% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-demand.R
\name{.tmb_build_map}
\alias{.tmb_build_map}
\title{Build TMB Map List}
\usage{
.tmb_build_map(has_k, n_re)
}
\arguments{
\item{has_k}{Logical, whether k is estimated.}

\item{n_re}{Integer, number of random effects.}
}
\value{
Named list for TMB map argument.
}
\description{
Build TMB Map List
}
\keyword{internal}
