% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-demand.R
\name{.tmb_build_tmb_data}
\alias{.tmb_build_tmb_data}
\title{Build TMB Data List}
\usage{
.tmb_build_tmb_data(prepared, design, equation, n_re)
}
\arguments{
\item{prepared}{Output from .tmb_prepare_data().}

\item{design}{Output from .tmb_build_design_matrices().}

\item{equation}{Character string, equation type.}

\item{n_re}{Integer, number of random effects.}
}
\value{
A list suitable for TMB::MakeADFun data argument.
}
\description{
Build TMB Data List
}
\keyword{internal}
