% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{ExtraF}
\alias{ExtraF}
\title{ExtraF}
\usage{
ExtraF(dat, equation = "hs", groups = NULL, verbose = FALSE)
}
\arguments{
\item{dat}{Long form data frame}

\item{equation}{"hs"}

\item{groups}{NULL for all. Character vector matching groups in id column}

\item{verbose}{If TRUE, prints all output including models}
}
\value{
List of results and models
}
\description{
Extra Sum of Squares F-test
}
\details{
One alpha better than individual alphas?
}
\author{
Brent Kaplan <bkaplan.ku@gmail.com>
}
