% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-price.R
\name{fit_cp_nls}
\alias{fit_cp_nls}
\title{Fit Nonlinear Cross-Price Model using NLS}
\usage{
fit_cp_nls(
  data,
  equation = c("exponentiated", "exponential"),
  start_vals = NULL,
  iter = 100,
  bounds = list(lower = c(qalone = 0.1, I = -2, beta = 1e-05), upper = c(qalone = NA, I =
    2, beta = 5)),
  fallback_to_nlsr = TRUE,
  return_all = TRUE
)
}
\arguments{
\item{data}{A data frame containing columns: x and y.}

\item{equation}{Type of equation ("exponential" or "exponentiated").}

\item{start_vals}{Optional named list of starting values.}

\item{iter}{Number of iterations for nls.multstart.}

\item{bounds}{List with 'lower' and 'upper' bounds for parameters.}

\item{fallback_to_nlsr}{Logical; if TRUE, try nlsr when others fail.}

\item{return_all}{Logical; if TRUE, returns a list with model metadata.}
}
\value{
Fitted model object or list with model metadata.
}
\description{
Fit Nonlinear Cross-Price Model using NLS
}
