% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{FitCurves}
\alias{FitCurves}
\title{FitCurves}
\usage{
FitCurves(dat, equation, k = NULL, remq0e = FALSE, replfree = NULL,
  rem0 = FALSE, plotting = FALSE)
}
\arguments{
\item{dat}{data frame (long form) of purchase task data.}

\item{equation}{Character vector of length one. Accepts either "hs" for Hursh and Silberberg (2008) or "koff" for Koffarnus, Franck, Stein, and Bickel (2015). If "hs" and the first price (x) is 0, it will be replaced by replfree.}

\item{k}{A numeric vector of length one. Reflects the range of consumption in log10 units. If none provided, k will be calculated based on the max/min of the entire sample}

\item{remq0e}{If TRUE, removes consumption and price where price == 0. Default value is FALSE}

\item{replfree}{Optionally replaces price == 0 with specified value. Note, if fitting using equation == "hs", and 0 is first price, 0 gets replaced by replfree. Default value is .01}

\item{rem0}{If TRUE, removes all 0s in consumption data prior to analysis. Default value is FALSE.}

\item{plotting}{If TRUE, removes all 0s in consumption data prior to analysis. Default value is FALSE.}
}
\value{
Data frame, fitting params and CI's/SE's
}
\description{
Analyzes purchase task data
}
\author{
Shawn Gilroy <shawn.gilroy@temple.edu> Brent Kaplan <bkaplan4@ku.edu>
}

