% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{FitCurves}
\alias{FitCurves}
\title{FitCurves}
\usage{
FitCurves(dat, equation, k, method = NULL, detailed = FALSE, xcol = "x",
  ycol = "y", idcol = "id", plotcurves = FALSE, vartext = NULL,
  plotdestination = NULL)
}
\arguments{
\item{dat}{data frame (long form) of purchase task data.}

\item{equation}{Character vector of length one. Accepts either "hs" for Hursh and Silberberg (2008) or "koff" for Koffarnus, Franck, Stein, and Bickel (2015).}

\item{k}{A numeric (or character) vector of length one. Reflects the range of consumption in log10 units. If none provided, k will be calculated based on the max/min of the entire sample + .5. If k = "ind", k will be calculated per individual using max/min + .5. If k = "fit", k will be a free parameter on an individual basis. If k = "range", k will be calculated based on the max/min of the entire sample + .5.}

\item{method}{Character vector of length one accepts either "Mean" or "Pooled". If not NULL (default), data will be aggregrated appropriately and analyzed in the specified way.}

\item{detailed}{If TRUE, output will be a 3 element list including (1) dataframe of results, (2) list of model objects, (3) list of individual dataframes used in fitting. Default value is FALSE, which returns only the dataframe of results.}

\item{xcol}{The column name that should be treated as "x" data}

\item{ycol}{The column name that should be treated as "y" data}

\item{idcol}{The column name that should be treated as dataset identifier}

\item{plotcurves}{Boolean whether to create individual plots. If TRUE, a "plots/" directory is created one level above working directory}

\item{vartext}{Character vector specifying indices to report on plots. Valid indices include "Q0d", "Alpha", "Intensity", "EV", "Pmaxe", "Omaxe", "Pmaxd", "Omaxd", "K", "Q0se", "Alphase", "R2", "AbsSS"}

\item{plotdestination}{Destination of plots}
}
\value{
If detailed == FALSE (default), a dataframe of results. If detailed == TRUE, a 3 element list consisting of (1) dataframe of results, (2) list of model objects, (3) list of individual dataframes used in fitting
}
\description{
Analyzes purchase task data
}
\author{
Brent Kaplan <bkaplan4@ku.edu> Shawn Gilroy <shawn.gilroy@temple.edu>
}
