% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{GetAnalyticPmax}
\alias{GetAnalyticPmax}
\title{Get pmax}
\usage{
GetAnalyticPmax(Alpha, K, Q0)
}
\arguments{
\item{Alpha}{alpha parameter}

\item{K}{k parameter ( > lower limit )}

\item{Q0}{Q0}
}
\value{
Numeric
}
\description{
...
}
\details{
...
}
\author{
Shawn Gilroy <sgilroy1@lsu.edu>
}
