% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{GetDescriptives}
\alias{GetDescriptives}
\title{Get Purchase Task Descriptive Summary}
\usage{
GetDescriptives(dat)
}
\arguments{
\item{dat}{Dataframe (long form)}
}
\value{
Dataframe with descriptive statistics
}
\description{
Calculates descriptive statistics from purchase task data.
}
\details{
Provides the following descriptive statistics from purchase task data at each price: mean consumption, standard deviation of consumption, proportion of 0 values, and number of NAs.
}
\examples{
GetDescriptives(apt)
}
\author{
Brent Kaplan <bkaplan4@ku.edu>
}

