% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{GetEmpirical}
\alias{GetEmpirical}
\title{GetEmpirical}
\usage{
GetEmpirical(dat)
}
\arguments{
\item{dat}{data frame (long form) of purchase task data.}
}
\value{
Data frame of empirical measures
}
\description{
Calculates empirical measures for purchase task data
}
\details{
Will calculate and return the following empirical measures: Intensity, BP0, BP1, Omax, and Pmax
}
\author{
Brent Kaplan <bkaplan.ku@gmail.com>
}
