% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{GetK}
\alias{GetK}
\title{Get K}
\usage{
GetK(dat)
}
\arguments{
\item{dat}{Dataframe (long form)}
}
\value{
Numeric
}
\description{
Calculates a k value by looking for the max/min consumption across entire dataset and adds .5 to that range
}
\details{
Will look for minimum greater zero
}
\examples{
GetK(apt)
}
\author{
Brent Kaplan <bkaplan4@ku.edu>
}

