% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{GetK}
\alias{GetK}
\title{Get K}
\usage{
GetK(dat, mnrange = TRUE)
}
\arguments{
\item{dat}{Dataframe (long form)}

\item{mnrange}{Boolean for whether k should be calculated based on the mean range + .5}
}
\value{
Numeric
}
\description{
Calculates a k value by looking for the max/min consumption across entire dataset and adds .5 to that range
}
\details{
Will look for maximum/minimum greater zero
}
\examples{
GetK(apt)
}
\author{
Brent Kaplan <bkaplan.ku@gmail.com>
}
