% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{GetSharedK}
\alias{GetSharedK}
\title{Get Shared K}
\usage{
GetSharedK(dat, equation, remq0e, replfree, rem0)
}
\arguments{
\item{dat}{Dataframe (longform)}

\item{equation}{Character vector. Accepts either "hs" or "koff"}

\item{remq0e}{If TRUE, removes consumption and price where price == 0.}

\item{replfree}{Optionally replaces price == 0 with specified value. Note, if fitting using equation == "hs", and 0 is first price, 0 gets replaced by replfree.}

\item{rem0}{If TRUE, removes all 0s in consumption data prior to analysis.}
}
\value{
Numeric value of shared k
}
\description{
Finds shared k among selected datasets using global regression
}
\details{
Uses global regression to fit a shared k among datasets. Assumes the dataset is in its final form. As of now, only to be used within FitCurves
}
\author{
Brent Kaplan <bkaplan4@ku.edu>
}

