% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{GetValsForSim}
\alias{GetValsForSim}
\title{Get Values for SimulateDemand}
\usage{
GetValsForSim(dat)
}
\arguments{
\item{dat}{Dataframe (long form)}
}
\value{
List of 3: setaparams, sdindex, x
}
\description{
Gets values used in SimulateDemand
}
\details{
Gets values used in SimulateDemand
}
\examples{
GetValsForSim(apt)
}
\author{
Brent Kaplan <bkaplan.ku@gmail.com>
}
