% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{glance.beezdemand_tmb}
\alias{glance.beezdemand_tmb}
\title{Glance at a beezdemand_tmb Model}
\usage{
\method{glance}{beezdemand_tmb}(x, ...)
}
\arguments{
\item{x}{A \code{beezdemand_tmb} object.}

\item{...}{Additional arguments.}
}
\value{
A one-row tibble of model-level statistics.
}
\description{
Glance at a beezdemand_tmb Model
}
