% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-methods.R
\name{glance.cp_model_lmer}
\alias{glance.cp_model_lmer}
\title{Get model summaries from a mixed-effects cross-price model}
\usage{
\method{glance}{cp_model_lmer}(x, ...)
}
\arguments{
\item{x}{A cp_model_lmer object.}

\item{...}{Additional arguments passed to broom.mixed::glance.}
}
\value{
A tibble with model summary statistics.
}
\description{
Get model summaries from a mixed-effects cross-price model
}
