% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{print.beezdemand_tmb}
\alias{print.beezdemand_tmb}
\title{Print Method for TMB Mixed-Effects Demand Model}
\usage{
\method{print}{beezdemand_tmb}(x, ...)
}
\arguments{
\item{x}{An object of class \code{beezdemand_tmb}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for TMB Mixed-Effects Demand Model
}
\examples{
\donttest{
data(apt)
fit <- fit_demand_tmb(apt, equation = "exponential", verbose = 0)
print(fit)
}

}
