% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-hurdle-methods.R
\name{print.summary.beezdemand_cp_hurdle}
\alias{print.summary.beezdemand_cp_hurdle}
\title{Print method for summary.beezdemand_cp_hurdle}
\usage{
\method{print}{summary.beezdemand_cp_hurdle}(x, digits = 4, n = Inf, ...)
}
\arguments{
\item{x}{A summary.beezdemand_cp_hurdle object}

\item{digits}{Number of significant digits to print}

\item{n}{Number of rows to print for any tables (unused for this class).}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for summary.beezdemand_cp_hurdle
}
