% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{print.summary.beezdemand_tmb}
\alias{print.summary.beezdemand_tmb}
\title{Print Method for TMB Model Summary}
\usage{
\method{print}{summary.beezdemand_tmb}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{summary.beezdemand_tmb} object.}

\item{digits}{Number of significant digits.}

\item{...}{Additional arguments.}
}
\description{
Print Method for TMB Model Summary
}
