% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic-methods.R
\name{print.summary.cp_unsystematic}
\alias{print.summary.cp_unsystematic}
\title{Print method for summary.cp_unsystematic objects}
\usage{
\method{print}{summary.cp_unsystematic}(x, ...)
}
\arguments{
\item{x}{A summary.cp_unsystematic object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for summary.cp_unsystematic objects
}
