% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{ranef.beezdemand_tmb}
\alias{ranef.beezdemand_tmb}
\title{Extract Random Effects from TMB Model}
\usage{
\method{ranef}{beezdemand_tmb}(object, ...)
}
\arguments{
\item{object}{A \code{beezdemand_tmb} object.}

\item{...}{Additional arguments.}
}
\value{
Data frame with subject-level random effects.
}
\description{
Extract Random Effects from TMB Model
}
