% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-hurdle-methods.R
\name{summary.beezdemand_cp_hurdle}
\alias{summary.beezdemand_cp_hurdle}
\title{Summary method for beezdemand_cp_hurdle}
\usage{
\method{summary}{beezdemand_cp_hurdle}(object, ...)
}
\arguments{
\item{object}{A beezdemand_cp_hurdle object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for beezdemand_cp_hurdle
}
