% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-demand.R
\name{summary.beezdemand_fixed}
\alias{summary.beezdemand_fixed}
\title{Summary Method for beezdemand_fixed}
\usage{
\method{summary}{beezdemand_fixed}(object, ...)
}
\arguments{
\item{object}{A beezdemand_fixed object}

\item{...}{Additional arguments (ignored)}
}
\value{
A summary.beezdemand_fixed object (inherits from beezdemand_summary)
}
\description{
Summary Method for beezdemand_fixed
}
