% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{summary.beezdemand_tmb}
\alias{summary.beezdemand_tmb}
\title{Summarize a TMB Mixed-Effects Demand Model Fit}
\usage{
\method{summary}{beezdemand_tmb}(object, report_space = c("natural", "log10", "internal"), ...)
}
\arguments{
\item{object}{An object of class \code{beezdemand_tmb}.}

\item{report_space}{Character. Reporting space for core demand parameters.
One of \code{"internal"}, \code{"natural"}, \code{"log10"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
An object of class \code{summary.beezdemand_tmb} (also inherits
from \code{beezdemand_summary}).
}
\description{
Summarize a TMB Mixed-Effects Demand Model Fit
}
\examples{
\donttest{
data(apt)
fit <- fit_demand_tmb(apt, equation = "exponential", verbose = 0)
summary(fit)
summary(fit, report_space = "log10")
}

}
