% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_apa}
\alias{theme_apa}
\title{APA Theme}
\usage{
theme_apa(plot.box = FALSE)
}
\arguments{
\item{plot.box}{Boolean for a box around the plot}
}
\value{
ggplot theme
}
\description{
APA theme for ggplot
}
\details{
Theme for ggplot graphics that closely align with APA formatting
}
\author{
Brent Kaplan <bkaplan.ku@gmail.com>
}
