% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp-hurdle-methods.R
\name{tidy.beezdemand_cp_hurdle}
\alias{tidy.beezdemand_cp_hurdle}
\title{Tidy method for beezdemand_cp_hurdle}
\usage{
\method{tidy}{beezdemand_cp_hurdle}(x, ...)
}
\arguments{
\item{x}{A beezdemand_cp_hurdle object}

\item{...}{Additional arguments (ignored)}
}
\value{
A data frame of model coefficients
}
\description{
Tidy method for beezdemand_cp_hurdle
}
