% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed-demand.R
\name{tidy.beezdemand_fixed}
\alias{tidy.beezdemand_fixed}
\title{Tidy Method for beezdemand_fixed}
\usage{
\method{tidy}{beezdemand_fixed}(x, ...)
}
\arguments{
\item{x}{A beezdemand_fixed object}

\item{...}{Additional arguments (ignored)}
}
\value{
A tibble of model coefficients with columns: id, term, estimate,
  std.error, statistic, p.value, component
}
\description{
Tidy Method for beezdemand_fixed
}
