% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-methods.R
\name{tidy.beezdemand_tmb}
\alias{tidy.beezdemand_tmb}
\title{Tidy a beezdemand_tmb Model}
\usage{
\method{tidy}{beezdemand_tmb}(x, report_space = c("natural", "log10", "internal"), ...)
}
\arguments{
\item{x}{A \code{beezdemand_tmb} object.}

\item{report_space}{Character. One of \code{"natural"}, \code{"log10"}, \code{"internal"}.}

\item{...}{Additional arguments.}
}
\value{
A tibble of model coefficients.
}
\description{
Tidy a beezdemand_tmb Model
}
