% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdle-demand.R
\name{validate_hurdle_data}
\alias{validate_hurdle_data}
\title{Validate Hurdle Demand Data}
\usage{
validate_hurdle_data(data, y_var, x_var, id_var)
}
\arguments{
\item{data}{A data frame.}

\item{y_var}{Character string, name of consumption variable.}

\item{x_var}{Character string, name of price variable.}

\item{id_var}{Character string, name of subject ID variable.}
}
\value{
A cleaned data frame ready for model fitting.
}
\description{
Internal function to validate and prepare data for hurdle demand model fitting.
}
\keyword{internal}
